+++ NEWS.rst
@@ -49,6 +49,8 @@ Changed
 
 Removed
 -------
+- Needs attention support `#93
+  <https://github.com/orontee/argos/issues/93>`_
 - Remove album sort choice from preferences dialog `#86
   <https://github.com/orontee/argos/issues/86>`_